feat: Move implicit formatters to one place
Custom exception handlers is already done with User micro types.
Implicit formatters for User, Project and Auth was moved
to formatters package.
Micro types formatters moved as well.
