+++ controllers/src/main/scala/cromwell/pipeline/controller/UserController.scala
@@ -3,7 +3,8 @@ package cromwell.pipeline.controller
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
-import cromwell.pipeline.datastorage.dto.user.{ PasswordUpdateRequest, UserUpdateRequest }
+import cromwell.pipeline.datastorage.dto.{ PasswordUpdateRequest, UserUpdateRequest }
+import cromwell.pipeline.datastorage.formatters.UserFormatters._
 import cromwell.pipeline.datastorage.utils.auth.AccessTokenContent
 import cromwell.pipeline.service.UserService
 import de.heikoseeberger.akkahttpplayjson.PlayJsonSupport._
@@ -12,7 +13,6 @@ import scala.concurrent.ExecutionContext
 import scala.util.{ Failure, Success }
 
 class UserController(userService: UserService)(implicit executionContext: ExecutionContext) {
-
   val route: AccessTokenContent => Route = accessToken =>
     path("users") {
       concat(
