+++ model/src/main/scala/cromwell/pipeline/model/wrapper/UserId.scala
@@ -12,7 +12,6 @@ object UserId extends Wrapped.Companion {
   type Wrapper = UserId
   type Error = String
   val pattern: String = "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
-  implicit lazy val userIdFormat: Format[UserId] = wrapperFormat
   override protected def create(value: String): UserId = new UserId(value)
   override protected def validate(value: String): ValidationResult[String] = Validated.cond(
     value.matches(pattern),
