+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dao/entry/ProjectEntry.scala
@@ -3,8 +3,8 @@ package cromwell.pipeline.datastorage.dao
 import com.github.tminglei.slickpg.PgEnumSupport
 import cromwell.pipeline.datastorage.Profile
 import cromwell.pipeline.datastorage.dao.entry.UserEntry
-import cromwell.pipeline.model.wrapper.UserId
 import cromwell.pipeline.datastorage.dto.{ Project, ProjectId, Repository, Visibility }
+import cromwell.pipeline.model.wrapper.UserId
 import slick.basic.Capability
 import slick.jdbc.{ JdbcType, PostgresProfile }
 
@@ -31,7 +31,9 @@ trait ProjectProfileWithEnumSupport extends PostgresProfile with PgEnumSupport {
   }
 }
 
-trait ProjectEntry { this: Profile with UserEntry with ProjectProfileWithEnumSupport =>
+trait ProjectEntry {
+  this: Profile with UserEntry with ProjectProfileWithEnumSupport =>
+
   import Implicits._
   import api._
