--- repositories/src/main/scala/cromwell/pipeline/datastorage/dto/File.scala
@@ -1,16 +1,0 @@
-package cromwell.pipeline.datastorage.dto
-
-import play.api.libs.functional.syntax._
-import play.api.libs.json.{ Format, JsPath }
-
-object File {
-  case class UpdateFileRequest(branch: String, content: String, commitMessage: String)
-
-  object UpdateFileRequest {
-    implicit val updateFileRequestFormat: Format[UpdateFileRequest] =
-      (JsPath \ "branch")
-        .format[String]
-        .and((JsPath \ "content").format[String])
-        .and((JsPath \ "commit_message").format[String])(UpdateFileRequest.apply, unlift(UpdateFileRequest.unapply))
-  }
-}
