+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/Project.scala
@@ -1,9 +1,7 @@
 package cromwell.pipeline.datastorage.dto
 
-import java.nio.file.{ Path, Paths }
+import java.nio.file.Path
 
-import play.api.libs.functional.syntax._
-import play.api.libs.json._
 import cromwell.pipeline.model.wrapper.UserId
 import slick.lifted.MappedTo
 
@@ -16,74 +14,26 @@ final case class Project(
   visibility: Visibility = Private
 ) {
   def withRepository(repositoryPath: Option[String]): Project =
-    this.copy(repository = repositoryPath.map(Repository(_)))
-}
-object Project {
-  implicit lazy val projectFormat: OFormat[Project] = Json.format[Project]
-  implicit lazy val projectWrites: Writes[Project] = (project: Project) =>
-    Json.obj(
-      "projectId" -> project.projectId.value,
-      "ownerId" -> project.ownerId,
-      "name" -> project.name,
-      "active" -> project.active,
-      "repository" -> project.repository,
-      "visibility" -> Visibility.toString(project.visibility),
-      "path" -> project.projectId.value
-    )
+    this.copy(repository = repositoryPath.map(Repository))
 }
 
 final case class ProjectId(value: String) extends MappedTo[String]
 
-object ProjectId {
-  implicit lazy val projectIdFormat: Format[ProjectId] = implicitly[Format[String]].inmap(ProjectId.apply, _.value)
-}
-
 final case class Repository(value: String) extends MappedTo[String]
-
-object Repository {
-  implicit lazy val repositoryFormat: Format[Repository] = implicitly[Format[String]].inmap(Repository.apply, _.value)
-}
 
 final case class ProjectAdditionRequest(name: String)
 
-object ProjectAdditionRequest {
-  implicit lazy val projectAdditionFormat: OFormat[ProjectAdditionRequest] = Json.format[ProjectAdditionRequest]
-}
-
 final case class ProjectDeleteRequest(projectId: ProjectId)
-
-object ProjectDeleteRequest {
-  implicit lazy val projectDeleteFormat: OFormat[ProjectDeleteRequest] = Json.format[ProjectDeleteRequest]
-}
 
 final case class ProjectUpdateRequest(projectId: ProjectId, name: String, repository: Option[Repository])
 
-object ProjectUpdateRequest {
-  implicit val updateRequestFormat: OFormat[ProjectUpdateRequest] = Json.format[ProjectUpdateRequest]
-}
+final case class ProjectResponse(projectId: ProjectId, name: String, active: Boolean)
 
 final case class Version(name: String, message: String, target: String, commit: Commit)
-object Version {
-  implicit val versionPlayFormat: OFormat[Version] = Json.format[Version]
-}
 
 final case class Commit(id: String)
-object Commit {
-  implicit val commitFormat: OFormat[Commit] = Json.format[Commit]
-}
 
 final case class ProjectFile(path: Path, content: String)
-
-object ProjectFile {
-  implicit object ProjectFileFormat extends Format[ProjectFile] {
-    override def reads(json: JsValue): JsResult[ProjectFile] =
-      JsSuccess(ProjectFile(Paths.get((json \ "path").as[String]), (json \ "content").as[String]))
-
-    override def writes(o: ProjectFile): JsValue = JsObject(
-      Seq("path" -> JsString(o.path.toString), "content" -> JsString(o.content))
-    )
-  }
-}
 
 sealed trait Visibility
 case object Private extends Visibility
@@ -91,9 +41,6 @@ case object Internal extends Visibility
 case object Public extends Visibility
 
 object Visibility {
-  implicit lazy val visibilityFormat: Format[Visibility] =
-    implicitly[Format[String]].inmap(Visibility.fromString, Visibility.toString)
-
   def fromString(s: String): Visibility = s match {
     case "private"  => Private
     case "internal" => Internal
@@ -111,17 +58,6 @@ object Visibility {
 
 final case class FileContent(content: String)
 
-object FileContent {
-  implicit lazy val validateFileRequestFormat: OFormat[FileContent] = Json.format[FileContent]
-}
-
 final case class ProjectUpdateFileRequest(project: Project, projectFile: ProjectFile, version: Option[Version])
 
-object ProjectUpdateFileRequest {
-  implicit lazy val projectUpdateFileRequestFormat: OFormat[ProjectUpdateFileRequest] =
-    ((JsPath \ "project").format[Project] ~ (JsPath \ "projectFile").format[ProjectFile] ~ (JsPath \ "version")
-      .formatNullable[Version])(
-      ProjectUpdateFileRequest.apply,
-      unlift(ProjectUpdateFileRequest.unapply)
-    )
-}
+final case class UpdateFileRequest(branch: String, content: String, commitMessage: String)
