--- repositories/src/main/scala/cromwell/pipeline/datastorage/dto/ProjectResponse.scala
@@ -1,9 +1,0 @@
-package cromwell.pipeline.datastorage.dto
-
-import play.api.libs.json.{ Json, OFormat }
-
-final case class ProjectResponse(projectId: ProjectId, name: String, active: Boolean)
-
-object ProjectResponse {
-  implicit lazy val findProjectResponse: OFormat[ProjectResponse] = Json.format[ProjectResponse]
-}
