+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/User.scala
@@ -1,7 +1,6 @@
 package cromwell.pipeline.datastorage.dto
 
-import cromwell.pipeline.model.wrapper.{ Name, UserEmail, UserId }
-import play.api.libs.json.{ Json, OFormat }
+import cromwell.pipeline.model.wrapper.{ Name, Password, UserEmail, UserId }
 import slick.lifted.MappedTo
 
 final case class User(
@@ -15,12 +14,26 @@ final case class User(
   active: Boolean = true
 )
 
-object User {
-  implicit lazy val userFormat: OFormat[User] = Json.format[User]
-}
-
 final case class ProfilePicture(value: Array[Byte]) extends MappedTo[Array[Byte]]
 
-object ProfilePicture {
-  implicit lazy val profilePictureFormat: OFormat[ProfilePicture] = Json.format[ProfilePicture]
+final case class UserUpdateRequest(email: UserEmail, firstName: Name, lastName: Name)
+
+final case class PasswordUpdateRequest(currentPassword: Password, newPassword: Password, repeatPassword: Password)
+
+final case class UserNoCredentials(
+  userId: UserId,
+  email: UserEmail,
+  firstName: Name,
+  lastName: Name,
+  active: Boolean
+)
+object UserNoCredentials {
+  def fromUser(user: User): UserNoCredentials =
+    UserNoCredentials(
+      userId = user.userId,
+      email = user.email,
+      firstName = user.firstName,
+      lastName = user.lastName,
+      active = user.active
+    )
 }
