--- repositories/src/main/scala/cromwell/pipeline/datastorage/dto/auth/SignUpRequest.scala
@@ -1,10 +1,0 @@
-package cromwell.pipeline.datastorage.dto.auth
-
-import cromwell.pipeline.model.wrapper.{ Name, Password, UserEmail }
-import play.api.libs.json.{ Json, OFormat }
-
-final case class SignUpRequest(email: UserEmail, password: Password, firstName: Name, lastName: Name)
-
-object SignUpRequest {
-  implicit val signUpRequestFormat: OFormat[SignUpRequest] = Json.format[SignUpRequest]
-}
