--- repositories/src/main/scala/cromwell/pipeline/datastorage/dto/user/UserUpdateRequest.scala
@@ -1,10 +1,0 @@
-package cromwell.pipeline.datastorage.dto.user
-
-import cromwell.pipeline.model.wrapper.{ Name, UserEmail }
-import play.api.libs.json.{ Json, OFormat }
-
-final case class UserUpdateRequest(email: UserEmail, firstName: Name, lastName: Name)
-
-object UserUpdateRequest {
-  implicit val updateRequestFormat: OFormat[UserUpdateRequest] = Json.format[UserUpdateRequest]
-}
