+++ repositories/src/test/scala/cromwell/pipeline/datastorage/dao/repository/auth/AuthUtilsTest.scala
@@ -2,7 +2,7 @@ package cromwell.pipeline.datastorage.dao.repository.auth
 
 import java.time.Instant
 
-import cromwell.pipeline.datastorage.dto.auth.AuthResponse
+import cromwell.pipeline.datastorage.dto.AuthResponse
 import cromwell.pipeline.datastorage.utils.auth.{ AccessTokenContent, AuthUtils, RefreshTokenContent }
 import cromwell.pipeline.model.wrapper.UserId
 import cromwell.pipeline.utils.{ AuthConfig, ExpirationTimeInSeconds }
@@ -10,6 +10,7 @@ import org.scalatest.{ Assertion, Matchers, WordSpec }
 import pdi.jwt.algorithms.JwtHmacAlgorithm
 import pdi.jwt.{ Jwt, JwtAlgorithm }
 import play.api.libs.json.Json
+import cromwell.pipeline.datastorage.formatters.AuthFormatters._
 
 class AuthUtilsTest extends WordSpec with Matchers {
